ChannelTags: simplify manage tags dialog (don't use plugin settings for this)
additionally move icons into shared file
